// src/generated/client/worldmonitor/market/v1/service_client.ts
export interface CryptoQuote {
  name: string;
  symbol: string;
  price: number;
  change: number;
  sparkline: number[];
}